port: macOS/arm64 build, curses shim with X11 frontend, NetHack tiles, crash fixes
- port/: curses shim, X11 frontend, NetHack tiles matched by name (mktiles.py)
- Makefile: srogue-x11 target; play.sh
- rogue.h: prototypes for variadic msg/addmsg and pointer-returning
  charge_str/ring_num (arm64 crashes), stdlib/unistd (wasm traps on void
  srand48/free), daemon -> sr_daemon (libc clash); RVIP declarations
- save.c: fclose instead of close(fileno) (lost the buffered tail); restore
  unlinks without endwin/fork under the shim; web guards (__EMSCRIPTEN__)
- state.c: 4-byte longs read into an int (64-bit long kept stale bytes)
- io.c: arrow keys under the shim; main.c: wc_mapwin, web autosave restore,
  no setuid/passwd on the web

+++ state.c
@@ -680,7 +680,7 @@ int
 rs_read_long(int inf, long *i)
 {
     unsigned char bytes[4];
-    long input;
+    int input;			/* port: 4 bytes on disk; a 64-bit long kept 4 stale bytes */
     unsigned char *buf = (unsigned char *) &input;
     
     rs_read(inf, &input, 4);
@@ -694,7 +694,7 @@ rs_read_long(int inf, long *i)
         buf = bytes;
     }
     
-    *i = *((long *) buf);
+    *i = *((int *) buf);
 
     return(READSTAT);
 }
@@ -722,7 +722,7 @@ int
 rs_read_ulong(int inf, unsigned long *i)
 {
     unsigned char bytes[4];
-    unsigned long input;
+    unsigned int input;		/* port: 4 bytes on disk (see rs_read_long) */
     unsigned char *buf = (unsigned char *) &input;
     
     rs_read(inf, &input, 4);
@@ -736,7 +736,7 @@ rs_read_ulong(int inf, unsigned long *i)
         buf = bytes;
     }
     
-    *i = *((unsigned long *) buf);
+    *i = *((unsigned int *) buf);
 
     return(READSTAT);
 }
